Resolved firebase integration, created notification models, introduced notification builder

+++ lib/data/requests_retry_service.dart
@@ -1,5 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:flutter_user_sdk/data/cache_repository.dart';
+import 'package:flutter_user_sdk/utils/connection_service.dart';
 import 'package:flutter_user_sdk/utils/extensions/request_options_serializer.dart';
 
 class RequestsRetryService {
@@ -12,10 +13,11 @@ class RequestsRetryService {
   void resendRequests() async {
     final cachedRequests = cacheRepository.getCachedRequests();
 
+    if (!ConnectionService.instance.isConnected) return;
     for (final element in cachedRequests) {
       final requestOption = RequestOptionsSerializer.fromJson(element.object);
 
-      await Dio().fetch<Response>(requestOption).then(
+      await Dio().fetch<dynamic>(requestOption).then(
         (response) {
           if (response.statusCode == 200) {
             cacheRepository.removeRequest(key: element.key);
